Completed CRUD API's & Exception Handlers for Sprint Items

+++ src/main/java/com/onePlace/onePlaceBackend/DAO/AppDAOImpl.java
@@ -1,10 +1,13 @@
 package com.onePlace.onePlaceBackend.DAO;
 
 import com.onePlace.onePlaceBackend.Entity.OneHomeSprintItem;
+import com.onePlace.onePlaceBackend.ExceptionHandler.oneHomeSprintItemsExceptionHandlers.SprintItemNotFoundException;
 import jakarta.persistence.EntityManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Repository
 public class AppDAOImpl implements AppDAO {
@@ -24,7 +27,16 @@ public class AppDAOImpl implements AppDAO {
 
     @Override
     public OneHomeSprintItem get(int id) {
-        return entityManager.find(OneHomeSprintItem.class, id);
+        OneHomeSprintItem sprintItem = entityManager.find(OneHomeSprintItem.class, id);
+        if (sprintItem == null) {
+            throw new SprintItemNotFoundException("Sprint Item Does Not Exists: ");
+        }
+        return sprintItem;
+    }
+
+    @Override
+    public List<OneHomeSprintItem> getAll() {
+        return entityManager.createQuery("SELECT s FROM OneHomeSprintItem s", OneHomeSprintItem.class).getResultList();
     }
 
     @Override
@@ -37,6 +49,9 @@ public class AppDAOImpl implements AppDAO {
     @Transactional
     public void delete(int id) {
         OneHomeSprintItem sprintItem = entityManager.find(OneHomeSprintItem.class, id);
+        if (sprintItem == null) {
+            throw new SprintItemNotFoundException("Sprint Item Does Not Exists: ");
+        }
         entityManager.remove(sprintItem);
     }
 }
